@@ -15,16 +15,26 @@
     "@types/jsonwebtoken": "^8.5.8",
     "@types/mongoose": "^5.7.36",
     "ts-node-dev": "^1.0.0-pre.57",
-    "typescript": "^3.9.7"
+    "typescript": "^4.2.4"
   },
   "dependencies": {
+    "@graphql-tools/stitch": "^8.6.9",
+    "@types/bcryptjs": "^2.4.2",
     "@types/json2csv": "^5.0.3",
+    "@types/uuid": "^8.3.4",
+    "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
     "dotenv": "^16.0.0",
     "express": "^4.17.1",
+    "express-graphql": "^0.12.0",
     "express-session": "^1.17.2",
+    "graphql": "^15.8.0",
+    "graphql-middleware": "^6.1.25",
+    "graphql-shield": "^7.5.0",
+    "graphql-tools": "^8.2.8",
     "json2csv": "^5.0.7",
     "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.10.0"
+    "mongoose": "^5.10.0",
+    "uuid": "^8.3.2"
   }
 }
