@@ -1,29 +1,20 @@
 import mongoose from 'mongoose';
-import { Utils } from '../utils/Utils';
-import { connSchema, IConn } from './connection';
-import { ITransaction, transactionSchema } from './transation';
 
-export enum Role {
-  CLIENT = 'CLIENT',
+export enum UserRoleType {
   ADMIN = 'ADMIN',
+  USER = 'USER',
 }
 
 export interface IUser {
   name: string;
-  lastName: string;
-  age: number;
-  bankBalance: number;
-  role: Role;
-  accountNumber?: string;
-  password?: string;
-  connections?: IConn[];
-  transacctions?: ITransaction[];
-}
-
-export interface IBasicUser {
-  name: string;
-  age: number;
-  accountNumber: string;
+  surname: string;
+  email: string;
+  password: string;
+  passwordConfirm: string;
+  resetPasswordToken: string;
+  role: UserRoleType;
+  status: boolean;
+  uuid: string;
 }
 
 interface userModelInterface extends mongoose.Model<UserDoc> {
@@ -32,14 +23,13 @@ interface userModelInterface extends mongoose.Model<UserDoc> {
 
 export interface UserDoc extends mongoose.Document {
   name: string;
-  lastName: string;
-  age: number;
-  bankBalance: number;
-  role: Role;
-  accountNumber: string;
+  surname: string;
+  email: string;
   password: string;
-  connections?: IConn[];
-  transacctions?: ITransaction[];
+  resetPasswordToken: string;
+  role: UserRoleType;
+  status: boolean;
+  uuid: string;
 }
 
 const userSchema = new mongoose.Schema({
@@ -47,31 +37,36 @@ const userSchema = new mongoose.Schema({
     type: String,
     required: true,
   },
-  lastName: {
+  surname: {
     type: String,
     required: true,
   },
-  age: {
-    type: Number,
-    required: true,
-  },
-  bankBalance: {
-    type: Number,
-    required: true,
-  },
-  role: {
-    type: Role,
-    required: true,
-  },
-  accountNumber: {
+  email: {
     type: String,
     required: true,
-    index: true,
   },
   password: {
     type: String,
     required: true,
   },
+  resetPasswordToken: {
+    type: String,
+    required: true,
+  },
+  role: {
+    type: UserRoleType,
+    required: true,
+  },
+  status: {
+    type: Boolean,
+    required: true,
+  },
+  uuid: {
+    type: String,
+    required: true,
+    index: true,
+  },
+  /*
   connections: {
     type: [connSchema],
     required: false,
@@ -79,25 +74,24 @@ const userSchema = new mongoose.Schema({
   transacctions: {
     type: [transactionSchema],
     required: false,
-  },
+  },*/
 });
 
 userSchema.statics.build = (attr: IUser) => {
-  attr.accountNumber = Math.floor(Math.random() * 10000000000) + '';
-  attr.password = Utils.generateCustomPassword();
+  // attr.password = Utils.generateCustomPassword();
   return new User(attr);
 };
 
 const User = mongoose.model<UserDoc, userModelInterface>('User', userSchema);
-const role = Role.CLIENT;
+// const role = UserRoleType.CLIENT;
 // const connections = new Array<IConn>();
-User.build({
+/*User.build({
   name: 'Edwin',
-  lastName: 'Quishpe',
+  su: 'Quishpe',
   age: 30,
   bankBalance: 100,
   role,
   // connections,
-});
+});*/
 
 export { User };
